text_detect/requirements.txt
pytesseract
torch
torchvision
Pillow
opencv-python
natsort
nltk
lmdb
PyQt5
six